add responseRequest method

+++ src/main/java/com/codethecode/courseregistersystem/entity/Course.java
@@ -36,7 +36,7 @@ public class Course {
     private String studentId;
 
     @Column(name = "cost")
-    private Long cost;
+    private Integer cost;
 
 
 }
